notification-srv-refactoring
