+++ .gitignore
@@ -1,18 +1,19 @@
-node_modules/
-npm-debug.log
-
-coverage/
-scripts/bin/
-
-.vscode
-.project
-.settings
-
-*.log
 *~
 *.d.ts
-lib/*.d.ts
-lib/*.js
 *.js
-!setupTopics.js
+*.log
+
+./test/*.js
+.directory
 .nyc_output
+.project
+.settings
+.vscode
+
+!setupTopics.js
+
+coverage/
+lib/
+node_modules/
+npm-debug.log
+scripts/bin/
