+++ .travis.yml
@@ -1,7 +1,7 @@
 language: node_js
 sudo: required
 node_js:
-    - "12.4.0"
+  - "12.4.0"
 git:
   depth: 10
 services:
