+++ package.json
@@ -2,7 +2,7 @@
   "name": "@restorecommerce/notification-srv",
   "version": "0.0.1",
   "description": "Restore Commerce notification microservice",
-  "main": "start.js",
+  "main": "lib/start.js",
   "author": "Invend GmbH",
   "repository": {
     "type": "git",
@@ -23,13 +23,20 @@
     "@restorecommerce/mailer": "^0.1.4",
     "@restorecommerce/protos": "github:restorecommerce/protos",
     "@restorecommerce/service-config": "^0.1.6",
-    "lodash": "^4.17.14"
+    "lodash": "^4.17.14",
+    "ts-node": "^8.5.4"
   },
   "devDependencies": {
     "@restorecommerce/grpc-client": "0.0.4",
     "@types/mocha": "^5.2.7",
+    "@types/node": "^12.12.11",
+    "@typescript-eslint/eslint-plugin": "^2.8.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^2.8.0",
+    "@typescript-eslint/parser": "^2.8.0",
     "coveralls": "^3.0.9",
     "cross-env": "^6.0.3",
+    "eslint": "^6.7.0",
+    "eslint-plugin-prefer-arrow-functions": "^3.0.1",
     "mocha": "^6.2.2",
     "npm-run-all": "^4.1.5",
     "nyc": "^14.1.1",
@@ -39,20 +46,20 @@
     "typescript": "^3.7.2"
   },
   "scripts": {
-    "start": "npm run createtopics && node start.js",
-    "dev": "cross-env NODE_ENV=development npm run createtopics && node start.js",
+    "start": "npm run createtopics && node lib/start.js",
+    "dev": "cross-env NODE_ENV=development npm run createtopics && nodemon --watch './src/**/*.ts' --exec 'ts-node' src/start.ts",
     "pretest": "npm run build && npm run tsctests && npm run createtopics",
     "test": "cross-env NODE_ENV=test npm run lint && nyc npm run mocha",
-    "lint": "tslint lib/*.ts *_test.ts service.ts start.ts interfaces.ts",
+    "lint": "eslint src --ext .ts",
     "mocha": "cross-env NODE_ENV=test mocha -R spec --timeout 60000 test/notification_test.js --stack-trace --exit",
     "coveralls": "nyc report --reporter=text-lcov | coveralls",
     "test-debug": "cross-env NODE_ENV=test mocha test/*.js --inspect-brk --exit",
     "createtopics": "node setupTopics.js io.restorecommerce.notification io.restorecommerce.command",
     "tsctests": "tsc -d -p tsconfig.test.json",
     "build:tsc": "tsc -d",
-    "build:clean": "rimraf service.js start.js interfaces.js lib/*.js",
+    "build:clean": "rimraf lib",
     "build": "npm-run-all lint build:clean build:tsc",
-    "healthcheck": "node healthcheck.js"
+    "healthcheck": "node lib/healthcheck.js"
   },
   "engines": {
     "node": ">= 9.2.0"
