+++ src/healthcheck.ts
@@ -11,7 +11,7 @@ grpcConfig['service'] = grpcConfig['services'][cfg.get('serviceNames:cis')];
 grpcConfig['protos'] = ['io/restorecommerce/commandinterface.proto'];
 grpcConfig['timeout'] = 3000;
 
-logger.silly("Connecting to: " + 'grpc://' + grpcConfig['addr']);
+logger.silly('Connecting to: ' + 'grpc://' + grpcConfig['addr']);
 
 const client = new grpcClient(grpcConfig, logger);
 const command = client.makeEndpoint('command', 'grpc://' + grpcConfig['addr']);
@@ -24,7 +24,7 @@ Object.keys(grpcConfig['services']).forEach((service) => {
       const value = new Buffer(JSON.stringify({service})).toString('base64');
       const fullPayload = {name: 'health_check', payload: {type_url: 'payload', value}};
 
-      logger.silly("Executing: " + JSON.stringify(fullPayload));
+      logger.silly('Executing: ' + JSON.stringify(fullPayload));
 
       const response = await command(fullPayload);
