+++ src/start.ts
@@ -6,4 +6,4 @@ import * as sconfig from '@restorecommerce/service-config';
 
 const cfg = sconfig(process.cwd());
 const service = new Cluster(cfg);
-service.run('./service.js');
+service.run('./lib/service.js');
