+++ tsconfig.json
@@ -1,17 +1,25 @@
 {
   "compilerOptions": {
-    "target": "es6",
+    "target": "ES6",
     "module": "commonjs",
     "moduleResolution": "node",
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
+    "noUnusedLocals": false,
+    "skipLibCheck": false,
+    "outDir": "lib",
+    "typeRoots": [
+      "node_modules/@types"
+    ]
   },
   "include": [
-     "./lib/**/*.ts", "./start.ts", "./service.ts", "./healthcheck.ts"
+    "./src/**/*.ts"
   ],
   "exclude": [
-    "node_modules",
-    "docs"
+    "**/*.test.ts",
+    "doc",
+    "lib",
+    "node_modules"
   ]
 }
