--- tslint.json
@@ -1,57 +1,0 @@
-{
-  "rules": {
-    "adjacent-overload-signatures": true,
-    "typedef": [true, "call-signature", "parameter", "member-variable-declaration"],
-    "object-literal-key-quotes": [true, "as-needed"],
-    "object-literal-shorthand": true,
-    "only-arrow-functions": [false],
-    "arrow-parens": false,
-    "class-name": true,
-    "comment-format": [
-      true,
-      "check-space"
-    ],
-    "indent": [
-      true,
-      "spaces"
-    ],
-    "one-line": [
-      true,
-      "check-open-brace",
-      "check-whitespace"
-    ],
-    "no-use-before-declare": true,
-    "no-var-keyword": true,
-    "semicolon": [true, "always"],
-    "whitespace": [
-      true,
-        "check-branch",
-        "check-decl",
-        "check-operator",
-        "check-separator",
-        "check-type",
-        "check-typecast"
-    ],
-    "typedef-whitespace":
-    [
-      true,
-      {
-        "call-signature": "nospace",
-        "index-signature": "nospace",
-        "parameter": "nospace",
-        "property-declaration": "nospace",
-        "variable-declaration": "nospace"
-      },
-      {
-        "call-signature": "onespace",
-        "index-signature": "onespace",
-        "parameter": "onespace",
-        "property-declaration": "onespace",
-        "variable-declaration": "onespace"
-      }
-    ],
-    "no-internal-module": true,
-    "no-trailing-whitespace": true,
-    "no-inferrable-types": true
-  }
-}
